Make common rule for terraform for aws, gcp and azure
